fix(products): cambiar tipo de campo en migracion

+++ database/migrations/2025_12_17_232800_create_products_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('image_path');
-            $table->float('price_for_kg');
-            $table->integer('stock_quantity');
+            $table->decimal('price_for_kg', 10, 2);
+            $table->decimal('stock_quantity', 10, 2);
             $table->timestamps();
         });
     }
